Iniciando o processo para gerar resultado de busca. Falta filtar por estado, cidade e categoria.

+++ app/src/main/java/com/example/apptcc/Entities/Assessment.java
@@ -4,7 +4,7 @@ public class Assessment {
 
     private String date;
     private String comment;
-    private int assessmentValue;
+    private float assessmentValue;
     private String keyAssessment;
     private String uidAssessment;
 
@@ -36,7 +36,7 @@ public class Assessment {
         this.comment = comment;
     }
 
-    public int getAssessmentValue() {
+    public float getAssessmentValue() {
         return assessmentValue;
     }
